+++ websocket/src/main/java/com/dhh/websocket/RxWebSocketUtil.java
@@ -4,6 +4,7 @@ import android.os.SystemClock;
 import android.support.v4.util.ArrayMap;
 import android.util.Log;
 
+import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
@@ -23,6 +24,7 @@ import rx.android.schedulers.AndroidSchedulers;
 import rx.functions.Action0;
 import rx.functions.Action1;
 import rx.functions.Func1;
+import rx.functions.Func2;
 import rx.schedulers.Schedulers;
 
 
@@ -41,6 +43,9 @@ public class RxWebSocketUtil {
     private Map<String, Observable<WebSocketInfo>> observableMap;
     private Map<String, WebSocket> webSocketMap;
     private boolean showLog;
+    private String logTag = "RxWebSocket";
+    private long interval = 1;
+    private TimeUnit reconnectIntervalTimeUnit = TimeUnit.SECONDS;
 
     private RxWebSocketUtil() {
         try {
@@ -63,6 +68,12 @@ public class RxWebSocketUtil {
         client = new OkHttpClient();
     }
 
+    /**
+     * please use {@link RxWebSocket} to instead of it
+     *
+     * @return
+     */
+    @Deprecated
     public static RxWebSocketUtil getInstance() {
         if (instance == null) {
             synchronized (RxWebSocketUtil.class) {
@@ -100,6 +111,17 @@ public class RxWebSocketUtil {
         this.showLog = showLog;
     }
 
+    public void setShowLog(boolean showLog, String logTag) {
+        setShowLog(showLog);
+        this.logTag = logTag;
+    }
+
+    public void setReconnectInterval(long interval, TimeUnit timeUnit) {
+        this.interval = interval;
+        this.reconnectIntervalTimeUnit = timeUnit;
+
+    }
+
     /**
      * @param url      ws://127.0.0.1:8080/websocket
      * @param timeout  The WebSocket will be reconnected after the specified time interval is not "onMessage",
@@ -114,7 +136,12 @@ public class RxWebSocketUtil {
             observable = Observable.create(new WebSocketOnSubscribe(url))
                     //自动重连
                     .timeout(timeout, timeUnit)
-                    .retry()
+                    .retry(new Func2<Integer, Throwable, Boolean>() {
+                        @Override
+                        public Boolean call(Integer integer, Throwable throwable) {
+                            return throwable instanceof IOException;
+                        }
+                    })
                     //共享
                     .doOnUnsubscribe(new Action0() {
                         @Override
@@ -122,7 +149,7 @@ public class RxWebSocketUtil {
                             observableMap.remove(url);
                             webSocketMap.remove(url);
                             if (showLog) {
-                                Log.d("RxWebSocketUtil", "注销");
+                                Log.d(logTag, "unsubscribe");
                             }
                         }
                     })
@@ -282,7 +309,12 @@ public class RxWebSocketUtil {
             if (webSocket != null) {
                 //降低重连频率
                 if (!"main".equals(Thread.currentThread().getName())) {
-                    SystemClock.sleep(2000);
+                    long ms = reconnectIntervalTimeUnit.toMillis(interval);
+                    if (ms == 0) {
+                        ms = 1000;
+                    }
+                    SystemClock.sleep(ms);
+                    subscriber.onNext(WebSocketInfo.createReconnect());
                 }
             }
             initWebSocket(subscriber);
@@ -293,7 +325,7 @@ public class RxWebSocketUtil {
                 @Override
                 public void onOpen(final WebSocket webSocket, Response response) {
                     if (showLog) {
-                        Log.d("RxWebSocketUtil", url + " --> onOpen");
+                        Log.d(logTag, url + " --> onOpen");
                     }
                     webSocketMap.put(url, webSocket);
                     if (!subscriber.isUnsubscribed()) {
@@ -318,7 +350,7 @@ public class RxWebSocketUtil {
                 @Override
                 public void onFailure(WebSocket webSocket, Throwable t, Response response) {
                     if (showLog) {
-                        Log.e("RxWebSocketUtil", t.toString() + webSocket.request().url().uri().getPath());
+                        Log.e(logTag, t.toString() + webSocket.request().url().uri().getPath());
                     }
                     if (!subscriber.isUnsubscribed()) {
                         subscriber.onError(t);
@@ -333,14 +365,14 @@ public class RxWebSocketUtil {
                 @Override
                 public void onClosed(WebSocket webSocket, int code, String reason) {
                     if (showLog) {
-                        Log.d("RxWebSocketUtil", url + " --> onClosed:code= " + code);
+                        Log.d(logTag, url + " --> onClosed:code= " + code);
                     }
                 }
             });
             subscriber.add(new MainThreadSubscription() {
                 @Override
                 protected void onUnsubscribe() {
-                    webSocket.close(3000, "手动关闭");
+                    webSocket.close(3000, "close WebSocket");
                 }
             });
         }