+++ websocket/src/main/java/com/dhh/websocket/WebSokcetAction1.java
@@ -9,13 +9,16 @@ import rx.functions.Action1;
 
 /**
  * Created by dhh on 2017/11/2.
+ * <p>
+ * please use {@link WebSocketSubscriber}
  */
-
+@Deprecated
 public abstract class WebSokcetAction1 implements Action1<WebSocketInfo> {
     @CallSuper
     @Override
     public void call(WebSocketInfo webSocketInfo) {
         if (webSocketInfo.isOnOpen()) {
+
             onOpen(webSocketInfo.getWebSocket());
         } else if (webSocketInfo.getString() != null) {
             onMessage(webSocketInfo.getString());