+++ ChangeLog.md
@@ -1,4 +1,8 @@
-
+### 2.0.0
+ - 增加了config属性，可以配置重连频率，是否打印日志及日志tag等等
+ - 重构了调用API，使用更加简洁
+ - 对外隐藏了RxWebSocketUtil，现在统一由RxWebSocket代理完成
+ - 增加了重连(onReconnect)、关闭(onClose)事件回调
 ### 1.5.0
  - fix: [#11](https://github.com/dhhAndroid/RxWebSocket/issues/11)
  - 修复WebSocket打开事件发送错误.